Refactor colour message JS with better naming
