+++ app/javascript/packs/application.js
@@ -12,4 +12,4 @@ Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
 
-require("tooltip")
+require("colourMessage")
